requirements.txt
jinja2>=3.1.0
jsonschema>=4.17.0
pyyaml>=6.0
click>=8.1.0
